+++ tests/Classifiers/OneVsRestTest.php
@@ -9,6 +9,7 @@ use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\Attributes\Test;
 use PHPUnit\Framework\Attributes\Group;
+use PHPUnit\Framework\Attributes\RunInSeparateProcess;
 use Rubix\ML\DataType;
 use Rubix\ML\EstimatorType;
 use Rubix\ML\Datasets\Unlabeled;
@@ -23,9 +24,7 @@ use Rubix\ML\Backends\Backend;
 use Rubix\ML\Backends\Serial;
 use Rubix\ML\Backends\Amp;
 use Rubix\ML\Backends\Swoole;
-use Rubix\ML\Specifications\SpecificationChain;
 use Rubix\ML\Specifications\ExtensionIsLoaded;
-use Rubix\ML\Specifications\SwooleExtensionIsLoaded;
 
 #[Group('Classifiers')]
 #[CoversClass(OneVsRest::class)]
@@ -76,12 +75,7 @@ class OneVsRestTest extends TestCase
             'backend' => $ampBackend,
         ];
 
-        if (
-            SpecificationChain::with([
-                new SwooleExtensionIsLoaded(),
-                new ExtensionIsLoaded('igbinary'),
-            ])->passes()
-        ) {
+        if (ExtensionIsLoaded::with('swoole')->passes()) {
             $swooleBackend = new Swoole();
 
             yield (string) $swooleBackend => [
@@ -156,6 +150,7 @@ class OneVsRestTest extends TestCase
 
     #[DataProvider('provideBackends')]
     #[Test]
+    #[RunInSeparateProcess]
     public function trainPredictProba(Backend $backend) : void
     {
         $this->backend = $backend;